--- api/src/paths/project/{projectId}/survey/{surveyId}/observations/list.test.ts
@@ -1,186 +1,0 @@
-import chai, { expect } from 'chai';
-import { describe } from 'mocha';
-import sinon from 'sinon';
-import sinonChai from 'sinon-chai';
-import * as list from './list';
-import * as db from '../../../../../../database/db';
-import * as observation_view_queries from '../../../../../../queries/observation/observation-view-queries';
-import SQL from 'sql-template-strings';
-
-chai.use(sinonChai);
-
-describe('getObservationsList', () => {
-  afterEach(() => {
-    sinon.restore();
-  });
-
-  const dbConnectionObj = {
-    systemUserId: () => {
-      return null;
-    },
-    open: async () => {
-      // do nothing
-    },
-    release: () => {
-      // do nothing
-    },
-    commit: async () => {
-      // do nothing
-    },
-    rollback: async () => {
-      // do nothing
-    },
-    query: async () => {
-      // do nothing
-    }
-  };
-
-  const sampleReq = {
-    keycloak_token: {},
-    params: {
-      projectId: 1,
-      surveyId: 1
-    }
-  } as any;
-
-  let actualResult: any = null;
-
-  const sampleRes = {
-    status: () => {
-      return {
-        json: (result: any) => {
-          actualResult = result;
-        }
-      };
-    }
-  };
-
-  it('should throw a 400 error when no project id path param', async () => {
-    sinon.stub(db, 'getDBConnection').returns({
-      ...dbConnectionObj,
-      systemUserId: () => {
-        return 20;
-      }
-    });
-
-    try {
-      const result = list.getObservationsList();
-
-      await result(
-        { ...sampleReq, params: { ...sampleReq.params, projectId: null } },
-        (null as unknown) as any,
-        (null as unknown) as any
-      );
-      expect.fail();
-    } catch (actualError) {
-      expect(actualError.status).to.equal(400);
-      expect(actualError.message).to.equal('Missing required path param `projectId`');
-    }
-  });
-
-  it('should throw a 400 error when no survey id path param', async () => {
-    sinon.stub(db, 'getDBConnection').returns({
-      ...dbConnectionObj,
-      systemUserId: () => {
-        return 20;
-      }
-    });
-
-    try {
-      const result = list.getObservationsList();
-
-      await result(
-        { ...sampleReq, params: { ...sampleReq.params, surveyId: null } },
-        (null as unknown) as any,
-        (null as unknown) as any
-      );
-      expect.fail();
-    } catch (actualError) {
-      expect(actualError.status).to.equal(400);
-      expect(actualError.message).to.equal('Missing required path param `surveyId`');
-    }
-  });
-
-  it('should throw a 400 error when no sql statement returned for block observations', async () => {
-    sinon.stub(db, 'getDBConnection').returns({
-      ...dbConnectionObj,
-      systemUserId: () => {
-        return 20;
-      }
-    });
-
-    sinon.stub(observation_view_queries, 'getBlockObservationListSQL').returns(null);
-
-    try {
-      const result = list.getObservationsList();
-
-      await result(sampleReq, (null as unknown) as any, (null as unknown) as any);
-      expect.fail();
-    } catch (actualError) {
-      expect(actualError.status).to.equal(400);
-      expect(actualError.message).to.equal('Failed to build SQL get statement');
-    }
-  });
-
-  it('should return the observations on success', async () => {
-    const blockObservations = {
-      id: 1,
-      b_id: 2,
-      observation_cnt: 3,
-      start_datetime: '2020/04/04',
-      end_datetime: '2020/05/05'
-    };
-
-    const mockQuery = sinon.stub();
-
-    mockQuery.resolves({ rows: [blockObservations] });
-
-    sinon.stub(db, 'getDBConnection').returns({
-      ...dbConnectionObj,
-      systemUserId: () => {
-        return 20;
-      },
-      query: mockQuery
-    });
-
-    sinon.stub(observation_view_queries, 'getBlockObservationListSQL').returns(SQL`some query`);
-
-    const result = list.getObservationsList();
-
-    await result(sampleReq, sampleRes as any, (null as unknown) as any);
-
-    expect(actualResult).to.eql({
-      blocks: [
-        {
-          id: 1,
-          block_id: 2,
-          number_of_observations: 3,
-          start_time: '2020/04/04',
-          end_time: '2020/05/05'
-        }
-      ]
-    });
-  });
-
-  it('should return null blocks when block observations response has no rows', async () => {
-    const mockQuery = sinon.stub();
-
-    mockQuery.resolves({ rows: null });
-
-    sinon.stub(db, 'getDBConnection').returns({
-      ...dbConnectionObj,
-      systemUserId: () => {
-        return 20;
-      },
-      query: mockQuery
-    });
-
-    sinon.stub(observation_view_queries, 'getBlockObservationListSQL').returns(SQL`some query`);
-
-    const result = list.getObservationsList();
-
-    await result(sampleReq, sampleRes as any, (null as unknown) as any);
-
-    expect(actualResult.blocks).to.be.null;
-  });
-});
