--- api/src/paths/project/{projectId}/survey/{surveyId}/template/{templateId}/getSignedUrl.test.ts
@@ -1,180 +1,0 @@
-import chai, { expect } from 'chai';
-import { describe } from 'mocha';
-import sinon from 'sinon';
-import sinonChai from 'sinon-chai';
-import * as get_signed_url from './getSignedUrl';
-import * as db from '../../../../../../../database/db';
-import * as survey_occurrence_queries from '../../../../../../../queries/survey/survey-occurrence-queries';
-import SQL from 'sql-template-strings';
-import * as file_utils from '../../../../../../../utils/file-utils';
-
-chai.use(sinonChai);
-
-describe('getSingleTemplateURL', () => {
-  afterEach(() => {
-    sinon.restore();
-  });
-
-  const dbConnectionObj = {
-    systemUserId: () => {
-      return null;
-    },
-    open: async () => {
-      // do nothing
-    },
-    release: () => {
-      // do nothing
-    },
-    commit: async () => {
-      // do nothing
-    },
-    rollback: async () => {
-      // do nothing
-    },
-    query: async () => {
-      // do nothing
-    }
-  };
-
-  const sampleReq = {
-    keycloak_token: {},
-    params: {
-      projectId: 1,
-      surveyId: 2,
-      templateId: 3
-    }
-  } as any;
-
-  let actualResult: any = null;
-
-  const sampleRes = {
-    status: () => {
-      return {
-        json: (result: any) => {
-          actualResult = result;
-        }
-      };
-    }
-  };
-
-  it('should throw an error when projectId is missing', async () => {
-    sinon.stub(db, 'getDBConnection').returns(dbConnectionObj);
-
-    try {
-      const result = get_signed_url.getSingleTemplateURL();
-
-      await result(
-        { ...sampleReq, params: { ...sampleReq.params, projectId: null } },
-        (null as unknown) as any,
-        (null as unknown) as any
-      );
-      expect.fail();
-    } catch (actualError) {
-      expect(actualError.status).to.equal(400);
-      expect(actualError.message).to.equal('Missing required path param `projectId`');
-    }
-  });
-
-  it('should throw an error when surveyId is missing', async () => {
-    sinon.stub(db, 'getDBConnection').returns(dbConnectionObj);
-
-    try {
-      const result = get_signed_url.getSingleTemplateURL();
-
-      await result(
-        { ...sampleReq, params: { ...sampleReq.params, surveyId: null } },
-        (null as unknown) as any,
-        (null as unknown) as any
-      );
-      expect.fail();
-    } catch (actualError) {
-      expect(actualError.status).to.equal(400);
-      expect(actualError.message).to.equal('Missing required path param `surveyId`');
-    }
-  });
-
-  it('should throw an error when templateId is missing', async () => {
-    sinon.stub(db, 'getDBConnection').returns(dbConnectionObj);
-
-    try {
-      const result = get_signed_url.getSingleTemplateURL();
-
-      await result(
-        { ...sampleReq, params: { ...sampleReq.params, templateId: null } },
-        (null as unknown) as any,
-        (null as unknown) as any
-      );
-      expect.fail();
-    } catch (actualError) {
-      expect(actualError.status).to.equal(400);
-      expect(actualError.message).to.equal('Missing required path param `templateId`');
-    }
-  });
-
-  it('should throw a 400 error when no sql statement returned', async () => {
-    sinon.stub(db, 'getDBConnection').returns({
-      ...dbConnectionObj,
-      systemUserId: () => {
-        return 20;
-      }
-    });
-
-    sinon.stub(survey_occurrence_queries, 'getSurveyTemplateOccurrenceSQL').returns(null);
-
-    try {
-      const result = get_signed_url.getSingleTemplateURL();
-
-      await result(sampleReq, (null as unknown) as any, (null as unknown) as any);
-      expect.fail();
-    } catch (actualError) {
-      expect(actualError.status).to.equal(400);
-      expect(actualError.message).to.equal('Failed to build SQL get statement');
-    }
-  });
-
-  it('should return null when getting signed url from S3 fails', async () => {
-    const mockQuery = sinon.stub();
-
-    mockQuery.resolves({ rows: [{ key: 's3Key' }] });
-
-    sinon.stub(db, 'getDBConnection').returns({
-      ...dbConnectionObj,
-      systemUserId: () => {
-        return 20;
-      },
-      query: mockQuery
-    });
-
-    sinon.stub(survey_occurrence_queries, 'getSurveyTemplateOccurrenceSQL').returns(SQL`some query`);
-    sinon.stub(file_utils, 'getS3SignedURL').resolves(null);
-
-    const result = get_signed_url.getSingleTemplateURL();
-
-    await result(sampleReq, sampleRes as any, (null as unknown) as any);
-
-    expect(actualResult).to.equal(null);
-  });
-
-  it('should return the signed url response on success', async () => {
-    const mockQuery = sinon.stub();
-
-    mockQuery.resolves({ rows: [{ key: 's3Key' }] });
-
-    sinon.stub(db, 'getDBConnection').returns({
-      ...dbConnectionObj,
-      systemUserId: () => {
-        return 20;
-      },
-      query: mockQuery
-    });
-
-    sinon.stub(survey_occurrence_queries, 'getSurveyTemplateOccurrenceSQL').returns(SQL`some query`);
-    sinon.stub(file_utils, 'getS3SignedURL').resolves('myurlsigned.com');
-
-    const result = get_signed_url.getSingleTemplateURL();
-
-    await result(sampleReq, sampleRes as any, (null as unknown) as any);
-
-    expect(actualResult).to.eql('myurlsigned.com');
-  });
-});
