+++ app/src/constants/i18n.ts
@@ -22,14 +22,6 @@ export const CreatePermitsI18N = {
     'An error has occurred while attempting to create your permits, please try again. If the error persists, please contact your system administrator.'
 };
 
-export const AddBlockObservationI18N = {
-  cancelTitle: 'Cancel Add Block Observation',
-  cancelText: 'Are you sure you want to cancel?',
-  createErrorTitle: 'Error Adding Block Observation',
-  createErrorText:
-    'An error has occurred while attempting to add your block observation, please try again. If the error persists, please contact your system administrator.'
-};
-
 export const UploadProjectAttachmentsI18N = {
   cancelTitle: 'Cancel Upload',
   cancelText: 'Are you sure you want to cancel?',
